Refactor flash storage to use traits
This paves the way for sharing flash between mesh stack and
other components in a firmware
